+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 * 💥 **\[core\]** Introduce `ObserverInfo` interface that is used from within the spy subsystem. This change
   effectively removes the `dispose()` method from the public interface of Observers accessed solely through
   the spy subsystem.
+* **\[core\]** Implement `equals()` and `hashCode()` on `ComponentInfoImpl`.
 
 ### [v0.30](https://github.com/arez/arez/tree/v0.30) (2017-11-29)
 [Full Changelog](https://github.com/arez/arez/compare/v0.29...v0.30)